requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dj-database-url==2.3.0
Django==5.1.2
django-cors-headers==4.4.0
djangorestframework==3.15.1
gunicorn==23.0.0
h11==0.14.0
idna==3.7
packaging==24.1
psycopg2-binary==2.9.10
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.32.0
whitenoise==6.7.0
